requirements.txt
absl-py==0.10.0
astor==0.8.1
certifi==2020.6.20
gast==0.4.0
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
importlib-metadata==2.0.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.1
numpy==1.19.2
Pillow==7.2.0
protobuf==3.13.0
scipy==1.5.2
six==1.15.0
tensorboard==1.14.0
tensorboardX==1.1
termcolor==1.1.0
torch==1.2.0
torchvision==0.4.0
tqdm==4.50.2
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.3.0
opencv-python==4.1.0.25
tensorflow==1.14.0
tensorflow-estimator==1.14.0
